Refactor repeated IP/FQDN schema definitions into helper function
