+++ nutanix/services/clustersv2/resource_nutanix_cluster_fetch_networks_v2.go
@@ -113,12 +113,7 @@ func nodeListNetworkingDetailsSchema() *schema.Schema {
 							"cvm_ip": {
 								Type:     schema.TypeList,
 								Computed: true,
-								Elem: &schema.Resource{
-									Schema: map[string]*schema.Schema{
-										"ipv4": SchemaForValuePrefixLengthResource(ipv4PrefixLengthDefaultValue),
-										"ipv6": SchemaForValuePrefixLengthResource(ipv6PrefixLengthDefaultValue),
-									},
-								},
+								Elem:     SchemaForIPList(false),
 							},
 							"uplink_list": {
 								Type:     schema.TypeList,
@@ -176,34 +171,19 @@ func nodeListSchema() *schema.Schema {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"ipv4": SchemaForValuePrefixLengthResource(ipv4PrefixLengthDefaultValue),
-							"ipv6": SchemaForValuePrefixLengthResource(ipv6PrefixLengthDefaultValue),
-						},
-					},
+					Elem:     SchemaForIPList(false),
 				},
 				"hypervisor_ip": {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"ipv4": SchemaForValuePrefixLengthResource(ipv4PrefixLengthDefaultValue),
-							"ipv6": SchemaForValuePrefixLengthResource(ipv6PrefixLengthDefaultValue),
-						},
-					},
+					Elem:     SchemaForIPList(false),
 				},
 				"ipmi_ip": {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"ipv4": SchemaForValuePrefixLengthResource(ipv4PrefixLengthDefaultValue),
-							"ipv6": SchemaForValuePrefixLengthResource(ipv6PrefixLengthDefaultValue),
-						},
-					},
+					Elem:     SchemaForIPList(false),
 				},
 				"digital_certificate_map_list": {
 					Type:     schema.TypeList,
